+++ src/QNEthernetClient.cpp
@@ -128,9 +128,11 @@ bool EthernetClient::connect(const ip_addr_t *ipaddr, uint16_t port, bool wait) 
     // NOTE: conn_ could be set to NULL somewhere during the yield
     while (conn_ != nullptr && !conn_->connected &&
            (sys_now() - t) < connTimeout_) {
+      yield();
+#if QNETHERNET_LOOP_AFTER_YIELD
       // NOTE: Call Ethernet loop in case an overridden yield() doesn't
       Ethernet.loop();
-      yield();
+#endif  // QNETHERNET_LOOP_AFTER_YIELD
     }
     if (conn_ == nullptr || !conn_->connected) {
       close();
@@ -313,9 +315,11 @@ void EthernetClient::close(bool wait) {
         // NOTE: conn_ could be set to NULL somewhere during the yield
         while (conn_ != nullptr && conn_->connected &&
                (sys_now() - t) < connTimeout_) {
+          yield();
+#if QNETHERNET_LOOP_AFTER_YIELD
           // NOTE: Call Ethernet loop in case an overridden yield() doesn't
           Ethernet.loop();
-          yield();
+#endif  // QNETHERNET_LOOP_AFTER_YIELD
         }
 #endif  // !LWIP_ALTCP
       }